requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
streamlit-extras
openai
faiss-cpu
wikipedia
duckduckgo-search
cohere
tiktoken
pinecone-client
pypdf

# uncomment to use huggingface llms
# huggingface-hub

# uncomment to use instructor embeddings
# InstructorEmbedding
# sentence-transformers
